chore: enforce consistent file names and move GlobHelp constant to jsonnet
