+++ internal/clihelpers/helpers.go
@@ -1,30 +1,5 @@
 package clihelpers
 
-const JsonnetGlobHelp = `Glob Syntax:
-
-    pattern:
-        { term }
-
-    term:
-        '*'         matches any sequence of non-separator characters
-        '**'        matches any sequence of characters
-        '?'         matches any single non-separator character
-        '[' [ '!' ] { character-range } ']'
-                    character class (must be non-empty)
-        '{' pattern-list '}'
-                    pattern alternatives
-        c           matches character c (c != '*', '**', '?', '\', '[', '{', '}')
-        '\' c       matches character c
-
-    character-range:
-        c           matches character c (c != '\\', '-', ']')
-        '\' c       matches character c
-        lo '-' hi   matches character c for lo <= c <= hi
-
-    pattern-list:
-        pattern { ',' pattern }
-                    comma-separated (without spaces) patterns`
-
 var (
 	BuildVersion = ""
 	BuildTime    = ""
